backend/feat/msil: enable driver by operator after checking onboarding docs
