+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Action/Dashboard/Operator/Driver.hs
@@ -1,16 +1,14 @@
-{-# OPTIONS_GHC -Wwarn=unused-imports #-}
-
 module Domain.Action.Dashboard.Operator.Driver (getDriverOperatorFetchHubRequests, postDriverOperatorRespondHubRequest, opsHubRequestLockKey) where
 
 import qualified API.Types.ProviderPlatform.Operator.Driver
-import Data.OpenApi (ToSchema)
 import Data.Time hiding (getCurrentTime)
 import Domain.Action.Dashboard.RideBooking.Driver
 import qualified Domain.Types.Merchant
 import Domain.Types.OperationHubRequests
 import qualified Environment
 import EulerHS.Prelude hiding (id)
-import Kernel.External.Encryption (decrypt, encrypt, getDbHash)
+import Kernel.Beam.Functions (runInReplica)
+import Kernel.External.Encryption (getDbHash)
 import qualified Kernel.Prelude
 import qualified Kernel.Storage.Hedis as Redis
 import Kernel.Types.APISuccess
@@ -18,11 +16,14 @@ import qualified Kernel.Types.Beckn.Context
 import Kernel.Types.Error
 import qualified Kernel.Types.Id
 import Kernel.Utils.Common
-import Servant
-import qualified Storage.Queries.DriverLicense as DLQuery
+import qualified SharedLogic.DriverOnboarding.Status as SStatus
+import SharedLogic.Merchant (findMerchantByShortId)
+import Storage.Cac.TransporterConfig (findByMerchantOpCityId)
+import qualified Storage.CachedQueries.Merchant.MerchantOperatingCity as CQMOC
 import qualified Storage.Queries.OperationHubRequests as SQOHR
 import qualified Storage.Queries.OperationHubRequestsExtra as SQOH
-import Tools.Auth
+import qualified Storage.Queries.Person as QPerson
+import qualified Storage.Queries.Vehicle as QVehicle
 
 getDriverOperatorFetchHubRequests :: (Kernel.Types.Id.ShortId Domain.Types.Merchant.Merchant -> Kernel.Types.Beckn.Context.City -> Kernel.Prelude.Maybe (Kernel.Prelude.UTCTime) -> Kernel.Prelude.Maybe (Kernel.Prelude.UTCTime) -> Kernel.Prelude.Maybe (API.Types.ProviderPlatform.Operator.Driver.RequestStatus) -> Kernel.Prelude.Maybe (API.Types.ProviderPlatform.Operator.Driver.RequestType) -> Kernel.Prelude.Maybe (Kernel.Prelude.Int) -> Kernel.Prelude.Maybe (Kernel.Prelude.Int) -> Kernel.Prelude.Maybe (Kernel.Prelude.Text) -> Kernel.Prelude.Maybe (Kernel.Prelude.Text) -> Environment.Flow API.Types.ProviderPlatform.Operator.Driver.OperationHubReqResp)
 getDriverOperatorFetchHubRequests _merchantShortId _opCity mbFrom mbTo mbStatus mbReqType mbLimit mbOffset mbDriverId mbMobileNumber = do
@@ -37,7 +38,7 @@ getDriverOperatorFetchHubRequests _merchantShortId _opCity mbFrom mbTo mbStatus 
   pure $ API.Types.ProviderPlatform.Operator.Driver.OperationHubReqResp {requests = map castHubRequests reqList}
 
 postDriverOperatorRespondHubRequest :: (Kernel.Types.Id.ShortId Domain.Types.Merchant.Merchant -> Kernel.Types.Beckn.Context.City -> API.Types.ProviderPlatform.Operator.Driver.RespondHubRequest -> Environment.Flow APISuccess)
-postDriverOperatorRespondHubRequest _merchantShortId _opCity req = do
+postDriverOperatorRespondHubRequest merchantShortId opCity req = do
   now <- getCurrentTime
   Redis.whenWithLockRedis (opsHubRequestLockKey req.operationHubRequestId) 60 $ do
     opHubReq <- SQOHR.findByPrimaryKey (Kernel.Types.Id.Id req.operationHubRequestId) >>= (fromMaybeM (InternalError "Invalid operation hub request id"))
@@ -45,8 +46,24 @@ postDriverOperatorRespondHubRequest _merchantShortId _opCity req = do
     when (opHubReq.driverId.getId == req.driverId) $ do
       void $ SQOHR.updateStatusWithDetails (castReqStatusToDomain req.status) (Just req.remarks) (Just now) (Just (Kernel.Types.Id.Id req.operatorId)) (Kernel.Types.Id.Id req.operationHubRequestId)
       when (req.status == API.Types.ProviderPlatform.Operator.Driver.APPROVED && opHubReq.requestType == ONBOARDING_INSPECTION) $ do
-        fork "enable driver after inspection" $
-          void $ postDriverEnable _merchantShortId _opCity (Kernel.Types.Id.Id req.driverId)
+        fork "enable driver after inspection" $ do
+          let personId = Kernel.Types.Id.Id req.driverId
+          merchant <- findMerchantByShortId merchantShortId
+          merchantOpCity <- CQMOC.findByMerchantIdAndCity merchant.id opCity >>= fromMaybeM (MerchantOperatingCityNotFound $ "merchantShortId: " <> merchantShortId.getShortId <> " ,city: " <> show opCity)
+          transporterConfig <- findByMerchantOpCityId merchantOpCity.id Nothing >>= fromMaybeM (TransporterConfigNotFound merchantOpCity.id.getId) -- (Just (DriverId (cast personId)))
+          vehicle <- runInReplica $ QVehicle.findById personId >>= fromMaybeM (VehicleDoesNotExist personId.getId)
+          person <- runInReplica $ QPerson.findById personId >>= fromMaybeM (PersonNotFound personId.getId)
+          let language = fromMaybe merchantOpCity.language person.language
+          driverDocuments <- SStatus.fetchDriverDocuments personId merchantOpCity transporterConfig language
+          processedVehicleDocuments <- SStatus.fetchProcessedVehicleDocuments personId merchantOpCity transporterConfig language (Just vehicle)
+          vehicleDoc <-
+            find (\doc -> doc.registrationNo == vehicle.registrationNo) processedVehicleDocuments
+              & fromMaybeM (InternalError $ "Vehicle doc not found for driverId " <> personId.getId) -- TODO remove this InternalError
+          let makeSelfieAadhaarPanMandatory = Nothing -- is it correct?
+          allVehicleDocsVerified <- SStatus.checkAllVehicleDocsVerified merchantOpCity.id vehicleDoc makeSelfieAadhaarPanMandatory
+          allDriverDocsVerified <- SStatus.checkAllDriverDocsVerified merchantOpCity.id driverDocuments vehicleDoc makeSelfieAadhaarPanMandatory
+          when (allVehicleDocsVerified && allDriverDocsVerified) $
+            void $ postDriverEnable merchantShortId opCity (Kernel.Types.Id.Id req.driverId)
   pure Success
 
 opsHubRequestLockKey :: Text -> Text
